Enemy Movement and Grid tiles
Improved Enemy movement, enemies don't move from tile to tile diagonally anymore.

Changed ground tiles.

+++ Assets/Scripts/Controllers/Healthbar.cs
@@ -6,12 +6,12 @@ using UnityEngine.UI;
 public class Healthbar : MonoBehaviour {
     public Slider slider;
 
-    public void setMaxHpValue(int value) { 
+    public void setMaxHpValue(float value) { 
         slider.maxValue = value;
         slider.value = value;
     }
 
-    public void setHp(int value) {
+    public void setHp(float value) {
         slider.value = value;
     }
 }
